tools: imager: refactor as a lib instead of main package
